% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bbox_poly}
\alias{bbox_poly}
\title{Create polygon from bbox}
\usage{
bbox_poly(bbox, crs = 4326)
}
\arguments{
\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}",}

\item{crs}{spatial projection of bounding box argument, defaults to epsg: 4326",}
}
\description{
Create polygon from bbox
}
