% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bboxxx}
\alias{bboxxx}
\title{Example bounding box used for testing functions}
\format{
A named vector with names \code{xmin}, \code{ymin}, \code{xmax}, \code{ymax}
}
\usage{
bboxxx
}
\description{
Example bounding box used for testing functions
}
\keyword{datasets}
