% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/zzz.R
=======
% Please edit documentation in R/helpers.R
>>>>>>> aaeade92dc4fe1d59717d170c02ab33af483f8d3
\name{chk_create}
\alias{chk_create}
\title{Check if folder exists and create if not}
\usage{
chk_create(path)
}
<<<<<<< HEAD
=======
\arguments{
\item{path}{path of folder to use as output, create if it does not already exist}
}
>>>>>>> aaeade92dc4fe1d59717d170c02ab33af483f8d3
\description{
Check if folder exists and create if not
}
