% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_marine_bioregions-35396c60.R
\name{dp_35396c60}
\alias{dp_35396c60}
\title{Federal Marine Bioregions}
\usage{
dp_35396c60(
  output,
  input = NULL,
  crs = 4326,
  bbox = NULL,
  timespan = NULL,
  ...
)
}
\arguments{
\item{output}{output folder for queried data. That folder will be organized into the different files that are loaded, i.e. raw data, formatted data, metadata and bibtex files. By default, data are stored in a root folder called 'data/'}

\item{input}{input folder in case data has to be loaded from disk, e.g. for datasets that are bound by a data sharing agreement. By default, raw data are searched for in a folder called 'data/data-raw/'}

\item{crs}{spatial projection to use to transform the spatial data into a uniform projection}

\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{timespan}{time span to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried \code{c(year1, year2, ...)}}

\item{...}{further arguments used in individual data pipelines, if applicable.}
}
\description{
Federal Marine Bioregions

The spatial planning framework for Canada's national network of Marine Protected Areas (MPAs) is comprised of 13 ecologically defined bioregions that cover Canada's oceans and the Great Lakes.
}
\examples{
\dontrun{
dp_35396c60()
}
}
\seealso{
\code{\link{pipedat}}
}
\concept{pipeline functions}
\keyword{35396c60}
\keyword{pipeline_id:}
