% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bbox_crop}
\alias{bbox_crop}
\alias{timespan_filter}
\alias{dp_parameters}
\title{Subset data queried using pipedat functions}
\usage{
bbox_crop(dat, bbox, bbox_crs, data_crs = sf::st_crs(dat))

timespan_filter(dat, timespan)

dp_parameters(
  dat,
  bbox = NULL,
  bbox_crs = NULL,
  data_crs = sf::st_crs(dat),
  timespan = NULL
)
}
\arguments{
\item{dat}{object to filter}

\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{data_crs}{spatial projection of object to filter, defaults to sf::st_crs(dat)"}

\item{timespan}{time span to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried \code{c(year1, year2, ...)}}

\item{ingrid}{logical, whether data should be integrated in a study grid, if available for specific pipeline. Default is set to \code{TRUE}.}

\item{keep_raw}{logical, whether raw data should be kept or removed. Default is set to \code{TRUE}. In any case, the raw data folder is compressed to save user disk space. Raw data can be accessed manually by unzipping \code{raw.zip}}
}
\description{
This function is used to subset the data based on a bounding box and/or a time span
}
\section{Functions}{
\itemize{
\item \code{bbox_crop}: crop object based on bounding box

\item \code{timespan_filter}: filter object by year

\item \code{dp_parameters}: crop object based on bounding box and filter based on year
}}

