% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griddat.R
\name{griddat}
\alias{griddat}
\title{Create spatial objects from grid and integrated data}
\usage{
griddat(uid, grid = NULL)
}
\arguments{
\item{uid}{unique identifiers of queried datasets. The full list of available data pipelines can be consulted using \code{pipelines()}}

\item{grid}{spatial grid used for data integration. Can be a \code{sf} object containing polygons or a \code{stars} rasters that will be used as a template. Individual cells must be identified by a unique identifier called \code{uid}. If NULL, the function attempts to load a grid \code{./data/data-grid/grid.poly.geojson}.}
}
\value{
This function returns objects in the global R environment
}
\description{
This function is used to join integrated data with study grid
}
\examples{
\dontrun{
importdat("f635934a")
}
}
