% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdat.R
\name{importdat}
\alias{importdat}
\title{Imports cleaned and integrated data in global R environment}
\usage{
importdat(uid)
}
\arguments{
\item{uid}{unique identifier of queried data. The full list of available data pipelines can be consulted using \code{pipelines()}}
}
\value{
This function returns objects in the global R environment
}
\description{
This function is used to import cleaned and integrated data
in the global R environment for use in other pipelines or
user processes.
}
\examples{
\dontrun{
importdat("f635934a")
}
}
