% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterload.R
\name{masterload}
\alias{masterload}
\title{Imports data in global R environment}
\usage{
masterload(path)
}
\arguments{
\item{path}{path to disk of the file to import}
}
\value{
This function returns objects in the global R environment
}
\description{
This function is used to import data of various format in R
}
\examples{
\dontrun{
masterload("path")
}
}
