% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{get_metadata}
\alias{add_metadata}
\title{Create a yaml metadata file}
\usage{
metadata(
  pipeline_id,
  pipeline_creators,
  pipeline_date,
  pipeline_url,
  pipeline_crs,
  pipeline_bbox = NULL,
  pipeline_timespan = NULL,
  data_name,
  data_description,
  data_access = timestamp(),
  data_timespan = NULL,
  data_bbox = NULL,
  data_contacts,
  data_url = NULL,
  data_uuid = NULL,
  data_availability = NULL,
  data_citekey
)

get_metadata(
  pipeline_id,
  pipeline_crs = 4326,
  pipeline_bbox = NULL,
  pipeline_timespan = NULL,
  data_access = timestamp(),
  data_bbox = NULL,
  data_timespan = NULL,
  ...
)

add_metadata(meta, ...)
}
\arguments{
\item{pipeline_id}{individual id of data pipeline}

\item{pipeline_creators}{creators of the data pipeline}

\item{pipeline_date}{date YYYY-MM-DD (\%Y-\%m-\%s) that pipeline was created}

\item{pipeline_url}{url to data pipeline code}

\item{pipeline_crs}{spatial projection used to transform the spatial data into a uniform projection. Default is set to \code{crs = 4326}}

\item{pipeline_bbox}{bounding box that was used to spatially subset the queried data, if applicable. The bounding box should be of the form c(xmin, ymin, xmax, ymax),}

\item{pipeline_timespan}{time span that was used to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried c(year1, year2, ...),}

\item{data_name}{short name of data queried}

\item{data_description}{description of data queried}

\item{data_access}{access date of queried data}

\item{data_timespan}{temporal coverage of data, if applicable}

\item{data_bbox}{bounding box of data of the form c(xmin, ymin, xmax, ymax), if applicable}

\item{data_contacts}{contacts for data queried}

\item{data_url}{data url, if applicable}

\item{data_uuid}{data uuid, e.g. from API or open data portals, if applicable}

\item{data_availability}{availability of queried data, one of c('open','on demand','data sharing agreement','restricted')}

\item{data_citekey}{vector of citation keys for reference to bibtex files}

\item{...}{further arguments used in \code{add_template()} to include additional metadata information on the queried data.}

\item{meta}{meta list generated by \code{metadata()} and used as first argument of \code{add_metadata()}}
}
\value{
This function returns a yaml metadata
}
\description{
This function is used to create a yaml metadata file when executing
a data pipeline that describes the data pipeline and the data queried
through the data pipeline.
}
\section{Functions}{
\itemize{
\item \code{metadata}: export metadata yaml

\item \code{get_metadata}: builds metadata from internal package data

\item \code{add_metadata}: add additional information on queried data to metadata
}}

\examples{
\dontrun{
meta <- get_metadata(
  pipeline_id = uid,
  pipeline_crs = crs,
  pipeline_bbox = bbox,
  pipeline_timespan = timespan,
  data_access = timestamp(),
  data_bbox = sf::st_bbox(dat),
)

meta <- add_metadata(meta,
  species = c("Cod", "Capelin", "Shrimp"),
  gear = c("Trawl", "Longline")
)
}
}
