% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipefigures.R
\name{figure_grid}
\alias{figure_grid}
\title{Function for basic plots}
\usage{
figure_grid(grid = NULL)
}
\arguments{
\item{grid}{sf object of the grid used for data integration. If NULL, the function attempts to
import \code{./data/data-grid/grid_poly.geojson}.}
}
\value{
Exports pngs of the figures selected
}
\description{
These functions are used to produce basic plots of the grid and data integrated using
the pipedat package, and used to produce a summary report using \code{pipereport()}.
}
\examples{
\dontrun{
figure_grid()
}
}
\concept{figure functions}
