% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipegrid.R
\name{pipegrid}
\alias{pipegrid}
\title{Creates a study grid}
\usage{
pipegrid()
}
\arguments{
\item{x}{soon described}

\item{y}{soon described}
}
\value{
soon described
}
\description{
This function is used to create a study grid for data integration and for a data workflow. It is a simple wrapper around \code{sf::st_make_grid}.
}
\examples{
\dontrun{
pipegrid()
}
}
