% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipein.R
\name{pipein}
\alias{pipein}
\title{Execute data integration pipelines}
\usage{
pipein(uid, output = "data", grid = NULL, ...)
}
\arguments{
\item{uid}{unique identifier of the data integration pipeline. The full list of available data pipelines can be consulted using \code{pipelines(type = "integrated")}}

\item{output}{output folder for integrated data. That folder will be organized into the different files that are loaded, i.e. raw data, formatted data, metadata and bibtex files. By default, data are stored in a root folder called 'data/'}

\item{grid}{spatial grid used for data integration. Can be a \code{sf} object containing polygons or a \code{stars} rasters that will be used as a template}

\item{...}{further arguments used in individual data integration pipelines, if applicable.}
}
\value{
This function returns the integrated data and associated metadata.
}
\description{
This function is used to format, analyze, integrate and incorporate data into a
user-specified study grid, if applicable, which we refer to as \emph{data integration pipelines}.
}
\examples{
\dontrun{
pipeint("0001")
}
}
