% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_paths}
\alias{make_paths}
\alias{check_files}
\alias{check_folders}
\alias{make_output}
\alias{get_pipeline}
\alias{get_shortname}
\alias{get_name}
\alias{get_description}
\alias{get_contact}
\alias{get_creator}
\alias{get_citekey}
\alias{get_bib}
\alias{get_pipeline_url}
\alias{get_pipeline_type}
\alias{get_rawid}
\alias{get_folderpaths}
\alias{make_filepaths}
\alias{get_filepaths}
\alias{get_grid_info}
\title{Series of functions to process the data and integration pipelines}
\usage{
make_paths(uid)

check_files(uid, ondisk = FALSE)

check_folders(uid)

make_output(uid)

get_pipeline(uid)

get_shortname(uid)

get_name(uid)

get_description(uid)

get_contact(uid)

get_creator(uid)

get_citekey(uid)

get_bib(uid)

get_pipeline_url(uid)

get_pipeline_type(uid)

get_rawid(uid)

get_folderpaths(uid)

make_filepaths(uid)

get_filepaths(uid)

get_grid_info(grd = NULL)
}
\arguments{
\item{uid}{unique identifier of queried data.}

\item{ondisk}{logical, whether the data should already be present locally, in cases when data are not available remotely (e.g. protected by data sharing agreements)}
}
\description{
Series of functions to process the data and integration pipelines
}
\section{Functions}{
\itemize{
\item \code{make_paths}: create path strings to folders and files

\item \code{check_files}: check if data is already present and send warning if data is already present and was thus not downloaded, or stop process if data needs to be available locally

\item \code{check_folders}: check if required folders exist

\item \code{make_output}: create required folders

\item \code{get_pipeline}: get pipeline information

\item \code{get_shortname}: get pipeline shortname

\item \code{get_name}: get pipeline name

\item \code{get_description}: get pipeline description

\item \code{get_contact}: get pipeline contacts

\item \code{get_creator}: get pipeline creator

\item \code{get_citekey}: get pipeline bibtex citekey

\item \code{get_bib}: get pipeline bibtex information

\item \code{get_pipeline_url}: get pipeline url

\item \code{get_pipeline_type}: get pipeline type

\item \code{get_rawid}: get raw data used for integration pipeline

\item \code{get_folderpaths}: get folder path to data

\item \code{make_filepaths}: create string of filepaths

\item \code{get_filepaths}: get path to files of a pipeline

\item \code{get_grid_info}: get information on a grid
}}

