% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeload.R
\name{pipeload}
\alias{pipeload}
\title{Execute data pipelines}
\usage{
pipeload(urls = NULL, govcan = NULL, output, large = FALSE)
}
\arguments{
\item{urls}{urls for data download}

\item{govcan}{unique identifier of resource on the open government federal data portal to download}

\item{output}{output folder for queried data.}

\item{large}{logical, whether file to download is large and thus R options should be modified for longer timeout options. Default set to false}
}
\value{
This function returns the queried raw data, formatted data, metadata and bibtex associated with the raw data.
}
\description{
This function is used to access, load and format a wide variety
of data; this function calls on a series of individual scripts
built to access data programmatically and reproducibly, which
we refer to as \emph{data pipelines}.
}
\examples{
\dontrun{
govcan <- "23eb8b56-dac8-4efc-be7c-b8fa11ba62e9"
output <- "data/"
if (file.exists(output)) dir.create(output)
pipeload(govcan = govcan, output = output)
}
}
