% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipenew.R
\name{pipenew}
\alias{pipenew}
\title{Creates a new data pipeline function}
\usage{
pipenew(name = NULL, template = "data_pipeline")
}
\arguments{
\item{name}{for a data pipeline or a data connect, name of the data that will be accessed through this new data pipeline. The name should be short and will only be used to ease the readability of the file structure, as each data pipeline is identified by a unique identifier rather than its name. For a data workflow, name of the data workflow; defaults to \code{data_workflow}}

\item{template}{name of the template to generate, one of \code{data_pipeline}, \code{data_connect}, \code{data_workflow}}
}
\value{
This function creates a new file \verb{R/dp_\{uid\}-\{name\}.R} from the template \code{inst/templates/data_pipeline.R}
}
\description{
This function is used to create a new data pipeline function from a template.
This function is meant to facilitate writing of data pipeline for the \code{pipedat}
developers and contributors
}
\examples{
\dontrun{
pipenew("fisheries")
}
}
