% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipereport.R
\name{pipereport}
\alias{pipereport}
\title{Create a report from the pipedat data in a project}
\usage{
pipereport()
}
\value{
This function creates a new report in \verb{pubs/pipedat/} using the template \code{inst/templates/report.Rmd} and all the data available in \verb{project-data/pipedat/} and \verb{project-data/grid/}.
}
\description{
This function creates a generic report that provides a summary of the data that was loaded and formatted throught the pipedat package in a specific location. The purpose of this report is to accompany a project report by providing the relevant information concerning the data that was accessed and formated through pipedat.
}
\examples{
\dontrun{
# Generate report
pipereport()
}
}
