% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeplot.R
\name{plot_legend_cont}
\alias{plot_legend_cont}
\alias{plot_legend_bin}
\title{Plot legend}
\usage{
plot_legend_cont(
  range = c(0, 1),
  pal = NULL,
  cexMain = 1,
  cexSub = 0.75,
  minUp = 0.175,
  mainTitle = NULL,
  subTitle = NULL,
  n = 5
)

plot_legend_bin(
  col,
  cexMain = 1,
  cexSub = 0.75,
  minUp = 0.175,
  mainTitle = NULL,
  subTitle = NULL
)
}
\arguments{
\item{range}{numeric, vector with minimal and maximal values}

\item{pal}{character, vector of colors, or color palette}

\item{cexMain}{numeric, cex for legend main text}

\item{cexSub}{numeric, cex for legend subtitle text}

\item{minUp}{numeric, minimum upper side to write as a function of bbox extent}

\item{mainTitle}{character, type of legend. Choices are 'continuous', 'binary' or 'categorical'}

\item{subTitle}{character, type of legend. Choices are 'continuous', 'binary' or 'categorical'}

\item{n}{numeric, number of ticks in the legend}

\item{col}{character, single color to use for binary data}
}
\value{
Opens a graphical interface with the plot
}
\description{
Function to create a legend
}
\keyword{legend}
\keyword{plot,}
