% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdat.R
\name{plotdat}
\alias{plotdat}
\alias{plotgrid}
\title{Function for basic plots}
\usage{
plotdat(uid)

plotgrid(grid = NULL)
}
\arguments{
\item{uid}{unique identifier of integrated data to plot}

\item{grid}{sf object of the grid used for data integration. If NULL, the function attempts to
import \code{./data/data-grid/grid_poly.geojson}.}
}
\value{
Exports pngs of the figures selected
}
\description{
These functions are used to produce basic plots of the grid and data integrated using
the pipedat package, and used to produce a summary report using \code{pipereport()}.
}
\examples{
\dontrun{
plotdat()
}
}
\concept{figure functions}
