% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trim_then_add}
\alias{trim_then_add}
\title{Remove and then add whitespaces again}
\usage{
trim_then_add(string)
}
\arguments{
\item{string}{character string to modify}
}
\description{
Remove and then add whitespaces again
}
