% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/zzz.R
\name{chk_create}
\alias{chk_create}
\title{Check if folder exists and create if not}
\usage{
chk_create(path)

chk_create(path)
}
\arguments{
\item{path}{path of folder to use as output, create if it does not already exist}
}
\description{
Check if folder exists and create if not

Check if folder exists and create if not
}
