% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/dp_marine_bioregions-35396c60.R,
%   R/dp_maritimes_grid-120a6032.R
\docType{data}
\name{data_pipelines}
\alias{data_pipelines}
\alias{dp_35396c60}
\alias{dp_120a6032}
\title{List of data pipelines}
\format{
A data frame:
\describe{
\item{pipeline_id}{unique identifier of data pipeline}
\item{name}{name of data pipeline}
\item{data_id}{unique identifier of the data obtained through the data pipeline}
}
}
\usage{
data_pipelines

dp_35396c60(output, name = NULL, crs = 4326, bbox = NULL, timespan = NULL, ...)

dp_120a6032(
  output,
  name = NULL,
  input = NULL,
  crs = 4326,
  bbox = NULL,
  timespan = NULL,
  ...
)
}
\arguments{
\item{output}{output folder for queried data. That folder will be organized into the different files that are loaded, i.e. raw data, formatted data, metadata and bibtex files. By default, data are stored in a root folder called 'data/'}

\item{crs}{spatial projection to use to transform the spatial data into a uniform projection}

\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{timespan}{time span to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried \code{c(year1, year2, ...)}}

\item{...}{further arguments used in individual data pipelines, if applicable.}
}
\description{
A dataset containing the list of data pipelines available in the \code{pipedat} packag
}
\examples{
\dontrun{
dp_35396c60()
}
\dontrun{
dp_120a6032()
}
}
\seealso{
\code{\link{pipedat}}

\code{\link{pipedat}}
}
\concept{pipeline functions}
\keyword{120a6032}
\keyword{35396c60}
\keyword{datasets}
\keyword{pipeline_id:}
