% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_3c82293e-fish.R, R/dp_bf3a7001-fish.R
\name{dp_3c82293e}
\alias{dp_3c82293e}
\alias{dp_bf3a7001}
\title{Shortname of dataset to be queried}
\usage{
dp_3c82293e(output, input = NULL, bbox = NULL, timespan = NULL, ...)

dp_bf3a7001(output, input = NULL, bbox = NULL, timespan = NULL, ...)
}
\arguments{
\item{output}{output folder for queried data. That folder will be organized into the different files that are loaded, i.e. raw data, formatted data, metadata and bibtex files. By default, data are stored in a root folder called 'data/'}

\item{input}{input folder in case data has to be loaded from disk, e.g. for datasets that are bound by a data sharing agreement. By default, raw data are searched for in a folder called 'data/data-raw/'}

\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{timespan}{time span to spatially subset the queried data, if applicable. The time span should be of the form \code{c(time_from, time_to)}}
}
\description{
Shortname of dataset to be queried

Short description of the dataset to be queried through this data pipeline

Shortname of dataset to be queried

Short description of the dataset to be queried through this data pipeline
}
\examples{
\dontrun{
dp_3c82293e()
}
\dontrun{
dp_bf3a7001()
}
}
\seealso{
\code{\link{pipedat}}

\code{\link{pipedat}}
}
\concept{pipeline functions}
\keyword{3c82293e}
\keyword{bf3a7001}
\keyword{dpid:}
