% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_cancoast-35608fef.R,
%   R/dp_contaminated_sites-804db12e.R,
%   R/dp_federal_marine_bioregions-f635934a.R,
%   R/dp_fisheries_logbooks-f2109e69.R, R/dp_marine_seismic-06230ea3.R,
%   R/dp_maritimes_grid-750b39f9.R
\name{dp_35608fef}
\alias{dp_35608fef}
\alias{dp_804db12e}
\alias{dp_f635934a}
\alias{dp_f2109e69}
\alias{dp_06230ea3}
\alias{dp_750b39f9}
\title{CanCoast - Coastal Materials Version 2.0}
\usage{
dp_35608fef(output, crs = 4326, bbox = NULL, timespan = NULL, ...)

dp_804db12e(output, crs = 4326, bbox = NULL, timespan = NULL, ...)

dp_f635934a(output, crs = 4326, bbox = NULL, timespan = NULL, ...)

dp_f2109e69(output, crs = 4326, bbox = NULL, timespan = NULL, ...)

dp_06230ea3(output, crs = 4326, bbox = NULL, timespan = NULL, ...)

dp_750b39f9(output, crs = 4326, bbox = NULL, timespan = NULL, ...)
}
\arguments{
\item{output}{output folder for queried data. That folder will be organized into the different files that are loaded, i.e. raw data, formatted data, metadata and bibtex files. By default, data are stored in a root folder called 'data/'}

\item{crs}{spatial projection to use to transform the spatial data into a uniform projection}

\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{timespan}{time span to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried \code{c(year1, year2, ...)}}

\item{...}{further arguments used in individual data pipelines, if applicable.}
}
\description{
CanCoast - Coastal Materials Version 2.0

CanCoast 2.0: data and indices to describe the sensitivity of Canada's marine coasts to changing climate

Federal Contaminated Sites Inventory

The Federal Contaminated Sites Inventory includes information on all known federal contaminated sites under the custodianship of departments, agencies and consolidated Crown corporations as well as those that are being or have been investigated to determine whether they have contamination arising from past use that could pose a risk to human health or the environment. The inventory also includes non-federal contaminated sites for which the Government of Canada has accepted some or all financial responsibility. It does not include sites where contamination has been caused by, and which are under the control of, enterprise Crown corporations, private individuals, firms or other levels of government.

Federal Marine Bioregions

The spatial planning framework for Canada's national network of Marine Protected Areas (MPAs) is comprised of 13 ecologically defined bioregions that cover Canada's oceans and the Great Lakes.

Commercial fisheries logbooks

A compilation of landing data from Zonal Interchange File Format (ZIFF) data between 2000 and 2020

Canadian National Marine Seismic Data Repository

The Geological Survey of Canada (Atlantic and Pacific) has collected marine survey field records on marine expeditions for over 50 years. This release makes available the results of an ongoing effort to scan and convert our inventory of analog marine survey field records (seismic, sidescan and sounder) to digital format.

Maritimes cumulative effects assessment study area grid

Gridded study area used jointly with N. Kelly and G. Murphy for the Maritimes region cumulative effects assessment
}
\examples{
\dontrun{
dp_35608fef()
}
\dontrun{
dp_804db12e()
}
\dontrun{
dp_f635934a()
}
\dontrun{
dp_f2109e69()
}
\dontrun{
dp_06230ea3()
}
\dontrun{
dp_750b39f9()
}
}
\seealso{
\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}
}
\concept{pipeline functions}
\keyword{06230ea3}
\keyword{35608fef}
\keyword{750b39f9}
\keyword{804db12e}
\keyword{f2109e69}
\keyword{f635934a}
\keyword{pipeline_id:}
