% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_federal_marine_bioregions-f635934a.R,
%   R/dp_maritimes_grid-750b39f9.R
\name{dp_f635934a}
\alias{dp_f635934a}
\alias{dp_750b39f9}
\title{Federal Marine Bioregions}
\usage{
dp_f635934a(output, crs = 4326, bbox = NULL, timespan = NULL, ...)

dp_750b39f9(output, crs = 4326, bbox = NULL, timespan = NULL, ...)
}
\arguments{
\item{output}{output folder for queried data. That folder will be organized into the different files that are loaded, i.e. raw data, formatted data, metadata and bibtex files. By default, data are stored in a root folder called 'data/'}

\item{crs}{spatial projection to use to transform the spatial data into a uniform projection}

\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{timespan}{time span to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried \code{c(year1, year2, ...)}}

\item{...}{further arguments used in individual data pipelines, if applicable.}
}
\description{
Federal Marine Bioregions

The spatial planning framework for Canada's national network of Marine Protected Areas (MPAs) is comprised of 13 ecologically defined bioregions that cover Canada's oceans and the Great Lakes.

Maritimes cumulative effects assessment study area grid

Gridded study area used jointly with N. Kelly and G. Murphy for the Maritimes region cumulative effects assessment
}
\examples{
\dontrun{
dp_f635934a()
}
\dontrun{
dp_750b39f9()
}
}
\seealso{
\code{\link{pipedat}}

\code{\link{pipedat}}
}
\concept{pipeline functions}
\keyword{750b39f9}
\keyword{f635934a}
\keyword{pipeline_id:}
