% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_fisheries_intensity-e2b7e6c4.R,
%   R/di_night_lights-aba5e90a.R, R/di_shipping_intensity-72312316.R
\name{di_e2b7e6c4}
\alias{di_e2b7e6c4}
\alias{di_aba5e90a}
\alias{di_72312316}
\title{Fisheries intensity}
\usage{
di_e2b7e6c4(grid = NULL, fishing_intensity_metric = 3, ...)

di_aba5e90a(grid = NULL, ...)

di_72312316(grid = NULL, shipping_type = "interpolated", ...)
}
\arguments{
\item{grid}{spatial grid used for data integration. Can be a \code{sf} object containing polygons or a \code{stars} rasters that will be used as a template}

\item{fishing_intensity_metric}{type of fishing intensity metric to evaluate, one of: 1: Fishing effort density, no normalization; 2: Fishing effort density; 3: Fishing biomass yield density; 4: Fishing relative biomass yield density. Passed on to eaMethods::fishing_intensity().}

\item{...}{further arguments used in individual data integration pipelines, if applicable.}

\item{shipping_type}{one of "interpolated" or "noninterpolated"}
}
\description{
Fisheries intensity

Assessment of the intensity of fishing activities divided by gear type as in Halpern et al. 2008 (i.e. demersal, destructive, high bycatch (DD); demersal, non-destructive, high bycatch (DNH); demersal, non-destructive, low bycatch (DNL); pelagic, high bycatch (PHB); pelagic, low bycatch (PLB)and gear mobility (i.e. fixed vs mobile).

Night lights

Assessment of the intensity of lights at night in coastal areas between 2012 and 2021

Shipping intensity

Assessment of the intensity of commercial shipping between 2017 and 2020 using data from the Global Fishing Watch.
}
\examples{
\dontrun{
di_e2b7e6c4()
}
\dontrun{
di_aba5e90a()
}
\dontrun{
di_72312316()
}
}
\seealso{
\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}
}
\concept{pipeline functions}
\keyword{72312316}
\keyword{aba5e90a}
\keyword{e2b7e6c4}
\keyword{pipeline_id:}
