% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_fisheries_intensity-e2b7e6c4.R
\name{di_e2b7e6c4}
\alias{di_e2b7e6c4}
\title{Fisheries intensity}
\usage{
di_e2b7e6c4(grid = NULL, fishing_intensity_metric = 3, ...)
}
\arguments{
\item{grid}{spatial grid used for data integration. Can be a \code{sf} object containing polygons or a \code{stars} rasters that will be used as a template}

\item{fishing_intensity_metric}{type of fishing intensity metric to evaluate, one of: 1: Fishing effort density, no normalization; 2: Fishing effort density; 3: Fishing biomass yield density; 4: Fishing relative biomass yield density. Passed on to eaMethods::fishing_intensity().}

\item{...}{further arguments used in individual data integration pipelines, if applicable.}
}
\description{
Fisheries intensity

Assessment of the intensity of fishing activities divided by gear type as in Halpern et al. 2008 (i.e. demersal, destructive, high bycatch (DD); demersal, non-destructive, high bycatch (DNH); demersal, non-destructive, low bycatch (DNL); pelagic, high bycatch (PHB); pelagic, low bycatch (PLB)and gear mobility (i.e. fixed vs mobile).
}
\examples{
\dontrun{
di_e2b7e6c4()
}
}
\seealso{
\code{\link{pipedat}}

Other pipeline functions: 
\code{\link{dp_68609420}()}
}
\concept{pipeline functions}
\keyword{e2b7e6c4}
\keyword{pipeline_id:}
