% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_direct_human_impact-99bb2d51.R,
%   R/di_fisheries_intensity-e2b7e6c4.R, R/di_halpern_cea-16c0d3ad.R,
%   R/di_marine_nammals_wwf_romm-aa3dcb26.R, R/di_night_lights-aba5e90a.R,
%   R/di_sea_surface_temperature_anomalies-606d08ea.R,
%   R/di_shipping_intensity-72312316.R
\name{di_99bb2d51}
\alias{di_99bb2d51}
\alias{di_e2b7e6c4}
\alias{di_16c0d3ad}
\alias{di_aa3dcb26}
\alias{di_aba5e90a}
\alias{di_606d08ea}
\alias{di_72312316}
\title{Direct human impact}
\usage{
di_99bb2d51(grid = NULL, ...)

di_e2b7e6c4(grid = NULL, fishing_intensity_metric = 3, ...)

di_16c0d3ad(grid = NULL, ...)

di_aa3dcb26(grid = NULL, ...)

di_aba5e90a(grid = NULL, ...)

di_606d08ea(grid = NULL, ...)

di_72312316(grid = NULL, shipping_type = "interpolated", ...)
}
\arguments{
\item{grid}{spatial grid used for data integration. Can be a \code{sf} object containing polygons or a \code{stars} rasters that will be used as a template}

\item{...}{further arguments used in individual data integration pipelines, if applicable.}

\item{fishing_intensity_metric}{type of fishing intensity metric to evaluate, one of: 1: Fishing effort density, no normalization; 2: Fishing effort density; 3: Fishing biomass yield density; 4: Fishing relative biomass yield density. Passed on to eaMethods::fishing_intensity().}

\item{shipping_type}{one of "interpolated" or "noninterpolated"}
}
\description{
Direct human impact

Assessment of the intensity of direct human impact along coastlines using population density as a proxy

Fisheries intensity

Assessment of the intensity of fishing activities divided by gear type as in Halpern et al. 2008 (i.e. demersal, destructive, high bycatch (DD); demersal, non-destructive, high bycatch (DNH); demersal, non-destructive, low bycatch (DNL); pelagic, high bycatch (PHB); pelagic, low bycatch (PLB)and gear mobility (i.e. fixed vs mobile).

Global cumulative human impacts assessments in 2008 and 2013

Integration of data from global cumulative human impacts assessments in 2008 and 2013 in study grid

Marine mammals

Assessment of the intensity of observations of marine mammals

Night lights

Assessment of the intensity of lights at night in coastal areas between 2012 and 2021

Sea surface temperature anomalies intensity

Assessment of the intensity of positive and negative sea surface temperature anomalies between 2013 and 2017.

Shipping intensity

Assessment of the intensity of commercial shipping between 2017 and 2020 using data from the Global Fishing Watch.
}
\examples{
\dontrun{
di_99bb2d51()
}
\dontrun{
di_e2b7e6c4()
}
\dontrun{
di_16c0d3ad()
}
\dontrun{
di_aa3dcb26()
}
\dontrun{
di_aba5e90a()
}
\dontrun{
di_606d08ea()
}
\dontrun{
di_72312316()
}
}
\seealso{
\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}
}
\concept{pipeline functions}
\keyword{16c0d3ad}
\keyword{606d08ea}
\keyword{72312316}
\keyword{99bb2d51}
\keyword{aa3dcb26}
\keyword{aba5e90a}
\keyword{e2b7e6c4}
\keyword{pipeline_id:}
