% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di_bathymetry_gebco_2021-ee0dd36a.R,
%   R/di_currents_climatology_bnam-62753cdf.R,
%   R/di_direct_human_impact-99bb2d51.R, R/di_fisheries_intensity-e2b7e6c4.R,
%   R/di_halpern_cea-16c0d3ad.R, R/di_invasive_species_richness-84b6ea0b.R,
%   R/di_marine_nammals_wwf_romm-aa3dcb26.R, R/di_night_lights-aba5e90a.R,
%   R/di_salinity_climatology_bnam-608c7f2f.R,
%   R/di_sea_surface_temperature_anomalies-606d08ea.R,
%   R/di_shipping_intensity-72312316.R,
%   R/di_species_list_nw_atlantic-893b37e8.R,
%   R/di_temperature_climatology_bnam-c055366b.R
\name{di_ee0dd36a}
\alias{di_ee0dd36a}
\alias{di_62753cdf}
\alias{di_99bb2d51}
\alias{di_e2b7e6c4}
\alias{di_16c0d3ad}
\alias{di_84b6ea0b}
\alias{di_aa3dcb26}
\alias{di_aba5e90a}
\alias{di_608c7f2f}
\alias{di_606d08ea}
\alias{di_72312316}
\alias{di_893b37e8}
\alias{di_c055366b}
\title{The GEBCO_2021 Grid}
\usage{
di_ee0dd36a(bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)

di_62753cdf(bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)

di_99bb2d51(bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)

di_e2b7e6c4(
  bbox = NULL,
  bbox_crs = NULL,
  timespan = NULL,
  grid = NULL,
  fishing_intensity_metric = 3,
  ...
)

di_16c0d3ad(bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)

di_84b6ea0b(bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)

di_aa3dcb26(bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)

di_aba5e90a(bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)

di_608c7f2f(bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)

di_606d08ea(bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)

di_72312316(
  bbox = NULL,
  bbox_crs = NULL,
  timespan = NULL,
  grid = NULL,
  shipping_type = "interpolated",
  ...
)

di_893b37e8(bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)

di_c055366b(bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)
}
\arguments{
\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{bbox_crs}{spatial projection of bounding box argument, defaults to epsg: 4326}

\item{timespan}{time span to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried \code{c(year1, year2, ...)}}

\item{grid}{spatial grid used for data integration. Can be a \code{sf} object containing polygons or a \code{stars} rasters that will be used as a template}

\item{...}{further arguments used in functions, if applicable.}

\item{fishing_intensity_metric}{type of fishing intensity metric to evaluate, one of: 1: Fishing effort density, no normalization; 2: Fishing effort density; 3: Fishing biomass yield density; 4: Fishing relative biomass yield density. Passed on to eaMethods::fishing_intensity().}

\item{shipping_type}{one of "interpolated" or "noninterpolated"}
}
\description{
The GEBCO_2021 Grid

The GEBCO_2021 Grid was published in July 2021 and is a global terrain model for ocean and land, providing elevation data, in meters, on a 15 arc-second interval grid

Monthly currents climatology

Monthly mean salinity from Bedford Institute of Oceanography North Atlantic Model (BNAM) results averaged over 1990 to 2015 period

Direct human impact

Assessment of the intensity of direct human impact along coastlines using population density as a proxy

Fisheries intensity

Assessment of the intensity of fishing activities divided by gear type as in Halpern et al. 2008 (i.e. demersal, destructive, high bycatch (DD); demersal, non-destructive, high bycatch (DNH); demersal, non-destructive, low bycatch (DNL); pelagic, high bycatch (PHB); pelagic, low bycatch (PLB)and gear mobility (i.e. fixed vs mobile).

Global cumulative human impacts assessments in 2008 and 2013

Integration of data from global cumulative human impacts assessments in 2008 and 2013 in study grid

Invasive species environmental stressor

Assessment of the risk associated with species invasions based on invasive species distribution models

Marine mammals

Assessment of the intensity of observations of marine mammals

Night lights

Assessment of the intensity of lights at night in coastal areas between 2012 and 2021

Monthly salinity climatology

Monthly mean salinity from Bedford Institute of Oceanography North Atlantic Model (BNAM) results averaged over 1990 to 2015 period

Sea surface temperature anomalies intensity

Assessment of the intensity of positive and negative sea surface temperature anomalies between 2013 and 2017.

Shipping intensity

Assessment of the intensity of commercial shipping between 2017 and 2020 using data from the Global Fishing Watch.

Species list North West Atlantic

Script to extract a list of species for the North West Atlantic using the CaRMS species checklist (id:084860fd) and observational data from various sources

Monthly temperature climatology

Monthly mean temperature from Bedford Institute of Oceanography North Atlantic Model (BNAM) results averaged over 1990 to 2015 period
}
\examples{
\dontrun{
di_ee0dd36a()
}
\dontrun{
di_62753cdf()
}
\dontrun{
di_99bb2d51()
}
\dontrun{
di_e2b7e6c4()
}
\dontrun{
di_16c0d3ad()
}
\dontrun{
di_84b6ea0b()
}
\dontrun{
di_aa3dcb26()
}
\dontrun{
di_aba5e90a()
}
\dontrun{
di_608c7f2f()
}
\dontrun{
di_606d08ea()
}
\dontrun{
di_72312316()
}
\dontrun{
di_893b37e8()
}
\dontrun{
di_c055366b()
}
}
\seealso{
\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

\code{\link{pipedat}}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}

Other pipeline functions: 
\code{\link{dp_68609420}()}
}
\concept{pipeline functions}
\keyword{16c0d3ad}
\keyword{606d08ea}
\keyword{608c7f2f}
\keyword{62753cdf}
\keyword{72312316}
\keyword{84b6ea0b}
\keyword{893b37e8}
\keyword{99bb2d51}
\keyword{aa3dcb26}
\keyword{aba5e90a}
\keyword{c055366b}
\keyword{e2b7e6c4}
\keyword{ee0dd36a}
\keyword{pipeline_id:}
