% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterload.R
\name{masterload}
\alias{masterload}
\alias{masterwrite}
\title{Imports and export data in or from global R environment}
\usage{
masterload(path)

masterwrite(obj, path)
}
\arguments{
\item{path}{path to disk of the file to import or export. Do not specify the extension if exporting using \code{masterwrite()}.}

\item{obj}{object to write to disk}
}
\value{
This function returns objects in the global R environment
}
\description{
These functions are used to import or export data of various format in R
}
\examples{
\dontrun{
masterload("path")
}
}
