% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{add_metadata}
\alias{people}
\title{Create a yaml metadata file}
\usage{
metadata(
  pipeline_id,
  pipeline_creators,
  pipeline_date,
  pipeline_url,
  data_name,
  data_pipeline_uuid,
  data_pipeline_crs,
  data_pipeline_bbox = NULL,
  data_pipeline_timespan = NULL,
  data_description,
  data_access = timestamp(),
  data_temporal = NULL,
  data_bbox = NULL,
  data_contacts,
  data_url = NULL,
  data_uuid = NULL,
  data_availability = NULL,
  data_citekey
)

add_metadata(meta, ...)

people(
  first_name = NULL,
  last_name = NULL,
  email = NULL,
  organization = NULL,
  department = NULL,
  role = NULL,
  developer = NULL
)
}
\arguments{
\item{pipeline_id}{individual id of data pipeline}

\item{pipeline_creators}{creators of the data pipeline, see \link{people}}

\item{pipeline_date}{date YYYY-MM-DD (\%Y-\%m-\%s) that pipeline was created}

\item{pipeline_url}{url to data pipeline code}

\item{data_name}{short name of data queried}

\item{data_pipeline_uuid}{uuid of the data associated with the pipeline. If no bbox or temporal constraits are used for the query, this uuid should always be the same. If a temporal range of a bounding box are imposed, this means that the resulting data are different from the original and thus a new uuid should be used.}

\item{data_pipeline_crs}{spatial projection used to transform the spatial data into a uniform projection,}

\item{data_pipeline_bbox}{bounding box that was used to spatially subset the queried data, if applicable. The bounding box should be of the form c(xmin, ymin, xmax, ymax),}

\item{data_pipeline_timespan}{time span that was used to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried c(year1, year2, ...),}

\item{data_description}{description of data queried}

\item{data_access}{access date of queried data}

\item{data_temporal}{temporal coverage of data, if applicable}

\item{data_bbox}{bounding box of data of the form c(xmin, ymin, xmax, ymax), if applicable}

\item{data_contacts}{contacts for data queried, see \link{people}}

\item{data_url}{data url, if applicable}

\item{data_uuid}{data uuid, e.g. from API or open data portals, if applicable}

\item{data_availability}{availability of queried data, one of c('open','on demand','data sharing agreement','restricted')}

\item{data_citekey}{vector of citation keys for reference to bibtex files}

\item{meta}{meta list generated by \code{metadata()} and used as first argument of \code{add_metadata()}}

\item{...}{further arguments used in \code{add_template()} to include additional metadata information on the queried data.}

\item{first_name}{first name}

\item{last_name}{last name}

\item{email}{email address}

\item{organization}{organization, affiliation}

\item{department}{department, branch, etc.}

\item{role}{role}

\item{developer}{predetermined "people", currently "david", "vincent"}
}
\value{
This function returns a yaml metadata
}
\description{
This function is used to create a yaml metadata file when executing
a data pipeline that describes the data pipeline and the data queried
through the data pipeline.
}
\section{Functions}{
\itemize{
\item \code{metadata}: export metadata yaml

\item \code{add_metadata}: add additional information on queried data to metadata

\item \code{people}: add pipeline creators and data contacts
}}

\examples{
\dontrun{
meta <- metadata(
  pipeline_id = "jhag76ad",
  pipeline_creators = people(developer = "david"),
  pipeline_date = "2022-04-13",
  pipeline_url = "https://path/to/pipeline.R",
  data_pipeline_uuid = "kjashf-aksjfg-soa8f7g-so8ef7",
  data_pipeline_crs = 4326,
  data_pipeline_bbox = c(xmin = -55, ymin = -55, xmax = -50, ymax = -50),
  data_pipeline_timespan = 2022,
  data_name = "Data X",
  data_description = "Data X characterizes Y",
  data_access = "2022-06-25",
  data_temporal = c(2000:2002, 2022),
  data_bbox = c(xmin = -60, ymin = -60, xmax = -40, ymax = -40),
  data_contacts = people("Jojo", "Le rigolo", "jojo@coolmain.com", "DFO", "Assistant"),
  data_url = "https://path/to/data/",
  data_uuid = "akiuy1-skjhfgkj-jhkasvf-mahsvbf",
  data_availability = "open",
  data_citekey = c("lerigolo2022", "archambault1990")
)

meta <- add_metadata(meta,
  species = c("Cod", "Capelin", "Shrimp"),
  gear = c("Trawl", "Longline")
)
}
}
