% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{get_metadata}
\alias{add_metadata}
\alias{gather_meta}
\alias{gather_bib}
\title{Create a yaml metadata file}
\usage{
metadata(
  pipeline_type,
  pipeline_id,
  pipeline_creators,
  pipeline_date,
  pipeline_url,
  pipeline_bbox = NULL,
  pipeline_timespan = NULL,
  name,
  description,
  contacts = NULL,
  access = timestamp(),
  citekey,
  data_timespan = NULL,
  data_bbox = NULL,
  data_url = NULL,
  data_uuid = NULL,
  data_availability = NULL,
  integration_data = NULL,
  integration_grid = NULL
)

get_metadata(
  pipeline_type,
  pipeline_id,
  pipeline_bbox = NULL,
  pipeline_timespan = NULL,
  access = timestamp(),
  data_bbox = NULL,
  data_timespan = NULL,
  integration_grid = NULL,
  ...
)

add_metadata(meta, ...)

add_metadata(meta, ...)

gather_meta()

gather_bib()
}
\arguments{
\item{pipeline_type}{type of pipeline, one of "data" or "integration"}

\item{pipeline_id}{individual id of data pipeline}

\item{pipeline_creators}{creators of the data pipeline}

\item{pipeline_date}{date YYYY-MM-DD (\%Y-\%m-\%s) that pipeline was created}

\item{pipeline_url}{url to data pipeline code}

\item{pipeline_bbox}{bounding box that was used to spatially subset the queried data, if applicable. The bounding box should be of the form c(xmin, ymin, xmax, ymax),}

\item{pipeline_timespan}{time span that was used to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried c(year1, year2, ...),}

\item{name}{short name of data queried}

\item{description}{description of data queried}

\item{contacts}{contacts for data queried}

\item{access}{access date of queried data}

\item{citekey}{vector of citation keys for reference to bibtex files}

\item{data_timespan}{temporal coverage of data, if applicable}

\item{data_bbox}{bounding box of data of the form c(xmin, ymin, xmax, ymax), if applicable}

\item{data_url}{data url, if applicable}

\item{data_uuid}{data uuid, e.g. from API or open data portals, if applicable}

\item{data_availability}{availability of queried data, one of c('open','on demand','data sharing agreement','restricted')}

\item{integration_data}{unique identifier of raw data pipelines used for the integration pipeline}

\item{integration_grid}{details of the grid used for data integration}

\item{...}{further arguments used in \code{add_template()} to include additional metadata information on the queried data.}

\item{meta}{meta list generated by \code{metadata()} and used as first argument of \code{add_metadata()}}

\item{grd}{object containing grid}
}
\value{
This function returns a yaml metadata
}
\description{
This function is used to create a yaml metadata file when executing
a data pipeline that describes the data pipeline and the data queried
through the data pipeline.
}
\section{Functions}{
\itemize{
\item \code{metadata}: export metadata yaml

\item \code{get_metadata}: builds metadata from internal package data

\item \code{add_metadata}: add additional information on queried data to metadata

\item \code{add_metadata}: add additional information on queried data to metadata

\item \code{gather_meta}: Script to extract metadata and contact from raw data list

\item \code{gather_bib}: Script to create a single bib file from the downloaded raw data using pipedat
}}

\examples{
\dontrun{
meta <- get_metadata(
  pipeline_id = uid,
  pipeline_bbox = bbox,
  pipeline_timespan = timespan,
  access = timestamp(),
  data_bbox = sf::st_bbox(dat),
)

meta <- add_metadata(meta,
  species = c("Cod", "Capelin", "Shrimp"),
  gear = c("Trawl", "Longline")
)
}
}
