% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipedat.R, R/zzz.R
\docType{package}
\name{pipedat}
\alias{pipedat}
\title{Execute data pipelines}
\usage{
pipedat(uid, crs = 4326, bbox = NULL, timespan = NULL, ...)
}
\arguments{
\item{uid}{unique identifier of queried data. The full list of available data pipelines can be consulted using \code{pipelines()}}

\item{crs}{spatial projection to use to transform the spatial data into a uniform projection}

\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{timespan}{time span to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried \code{c(year1, year2, ...)}}

\item{...}{further arguments used in individual data pipelines, if applicable.}
}
\value{
This function returns the queried raw data, formatted data, metadata and bibtex associated with the raw data.
}
\description{
This function is used to access, load and format a wide variety
of data; this function calls on a series of individual scripts
built to access data programmatically and reproducibly, which
we refer to as \emph{data pipelines}.
}
\examples{
\dontrun{
pipedat("0001")
}
}
