% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipedat.R
\name{pipedat}
\alias{pipedat}
\title{Execute data pipelines}
\usage{
pipedat(uid, output = NULL, bbox = NULL, timespan = NULL)
}
\arguments{
\item{uid}{unique identifier of queried data. The full list of available data pipelines can be consulted using \code{pipelines()}}

\item{output}{output folder for queried data. That folder will be organized into the different files that are loaded, i.e. raw data, formatted data, metadata and bibtex files.}

\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{timespan}{time span to spatially subset the queried data, if applicable. The time span should be of the form \code{c(time_from, time_to)}}
}
\value{
This function returns the queried raw data, formatted data, metadata and bibtex associated with the raw data.
}
\description{
This function is used to access, load and format a wide variety
of data; this function calls on a series of individual scripts
built to access data programmatically and reproducibly, which
we refer to as \emph{data pipelines}.
}
\examples{
\dontrun{
pipedat("0001")
}
}
