% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipedat.R, R/zzz.R
\docType{package}
\name{pipedat}
\alias{pipedat}
\title{Execute data pipelines}
\usage{
pipedat(uid, bbox = NULL, bbox_crs = NULL, timespan = NULL, grid = NULL, ...)
}
\arguments{
\item{uid}{unique identifier of queried data. The full list of available data pipelines can be consulted using \code{pipelines()}}

\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{bbox_crs}{spatial projection of bounding box argument, defaults to epsg: 4326}

\item{timespan}{time span to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried \code{c(year1, year2, ...)}}

\item{grid}{spatial grid used for data integration. Can be a \code{sf} object containing polygons or a \code{stars} rasters that will be used as a template}

\item{...}{further arguments used in functions, if applicable.}
}
\value{
This function returns the queried raw data, formatted data, metadata and bibtex associated with the raw data.
}
\description{
This function is used to access, load and format a wide variety
of data; this function calls on a series of individual scripts
built to access data programmatically and reproducibly, which
we refer to as \emph{data pipelines}.
}
\examples{
\dontrun{
pipedat("0001")
}
}
