% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipedat.R, R/zzz.R
\docType{package}
\name{pipedat}
\alias{pipedat}
\title{Execute data pipelines}
\usage{
pipedat(
  uid,
  bbox = NULL,
  timespan = NULL,
  integrate = TRUE,
  grd = "data/grid/grid.tif",
  keep_raw = TRUE,
  ...
)
}
\arguments{
\item{uid}{unique identifier of queried data. The full list of available data pipelines can be consulted using \code{pipelines()}}

\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{timespan}{time span to temporally subset the queried data, if applicable. The time span should a vector containing all the years to be queried \code{c(year1, year2, ...)}}

\item{integrate}{logical, whether data should be integrated in a study grid, if available for specific pipeline. Default is set to \code{TRUE}.}

\item{grd}{raster grid used for data integration, either a path to a \code{.tif} on disk or a \code{stars} objects. Default is set to \code{data/grid/grid.tif}, which is generated by \code{pipegrid()} or \code{pipeflow()}}

\item{keep_raw}{logical, whether raw data should be kept or removed. Default is set to \code{TRUE}. In any case, the raw data folder is compressed to save user disk space. Raw data can be accessed manually by unzipping \code{raw.zip}}

\item{...}{further arguments used in functions, if applicable.}
}
\value{
This function returns the queried raw data, formatted data, metadata and bibtex associated with the raw data.
}
\description{
This function is used to access, load and format a wide variety
of data; this function calls on a series of individual scripts
built to access data programmatically and reproducibly, which
we refer to as \emph{data pipelines}.
}
\examples{
\dontrun{
pipedat("0001")
}
}
