% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipegrid.R
\name{pipegrid}
\alias{pipegrid}
\title{Creates a raster study grid}
\usage{
pipegrid(aoi, cellsize, crs = 4326)
}
\arguments{
\item{cellsize}{target cellsize}

\item{crs}{object of class crs; coordinate reference system of the target of the target grid in case argument x is missing, if x is not missing, its crs is inherited.}

\item{x}{object of class sf or sfc}

\item{bbox}{bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form \code{c(xmin, ymin, xmax, ymax)}}
}
\value{
a tif file and a csv file
}
\description{
This function is used to create a raster study grid that will serve as a template for data extractions
}
\examples{
\dontrun{
cellsize <- 1
crs <- 4326
bb <- c(xmin = -45, ymin = -45, xmax = 45, ymax = 45)
bbox <- sf::st_bbox(bb, crs = crs)
aoi <- sf::st_as_sfc(bbox, crs = crs)
pipegrid(bb, cellsize, crs)
pipegrid(bbox, cellsize, crs)
pipegrid(aoi, cellsize, crs)
}
}
