% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelist.R
\name{pipelist}
\alias{pipelist}
\title{List of available data pipalines}
\usage{
pipelist(type = "data")
}
\arguments{
\item{type}{type of pipeline, one of \code{data} or \code{connect}}
}
\value{
This function prints a list of available data pipelines in the \code{pipedat} package
}
\description{
This function provides the list of individual scripts
built to access data programmatically and reproducibly, which
we refer to as \emph{data pipelines}.
}
\examples{
\dontrun{
pipelist()
}
}
