% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeplot.R
\name{pipeplot}
\alias{pipeplot}
\alias{pipeplot.sf}
\title{plot pipedat data maps}
\usage{
pipeplot(dat, ...)

\method{pipeplot}{sf}(
  dat,
  bbox,
  main = NULL,
  type = NULL,
  subtitle = NULL,
  unit_data = NULL,
  references = NULL,
  ...
)
}
\arguments{
\item{dat}{object of class sf}

\item{...}{further specifications, see \link{plot} and details.}

\item{bbox}{bounding box of the plot, should be of the form \code{c(xmin, ymin, xmax, ymax)}}

\item{main}{main title}

\item{type}{secondary title / data type (see metadata files)}

\item{subtitle}{subtitle}

\item{unit_data}{units of data}

\item{references}{data citation used for integrated data}
}
\description{
base plot functions for pipedat package
}
